requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiolimiter==1.1.0
aiosignal==1.3.1
amqp==5.2.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
asyncio==3.4.3
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
billiard==4.2.0
bleach==6.1.0
camelot-py==0.9.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cobble==0.1.4
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
configparser==3.5.3
contourpy==1.2.1
cryptography==43.0.0
cssselect==1.2.0
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
docx==0.2.4
elsapy==0.5.1
et-xmlfile==1.1.0
eutils==0.6.0
excalibur-py==0.4.3
executing==2.0.1
fastjsonschema==2.20.0
Flask==1.1.2
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
ghostscript==0.7
greenlet==3.0.3
h11==0.14.0
habanero==1.2.6
httpcore==1.0.5
httpx==0.27.0
humanfriendly==10.0
idna==3.7
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
kombu==5.3.7
Levenshtein==0.25.1
lxml==5.2.2
lxml_html_clean==0.2.0
mammoth==1.8.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
matplotlib-venn==1.1.1
metapub==0.5.12
mistune==3.0.2
multidict==6.0.5
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.1
notebook_shim==0.2.4
numpy==2.0.0
opencv-python==4.10.0.84
openpyxl==3.1.4
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pdfminer.six==20240706
pillow==10.3.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psutil==6.0.0
PuLP==2.9.0
pure-eval==0.2.2
pybliometrics==4.1
pycparser==2.22
Pygments==2.18.0
pyparsing==3.1.2
pypdf==4.3.1
PyPDF2==2.12.1
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-Levenshtein==0.25.1
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
rapidfuzz==3.9.5
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rispy==0.9.0
rpds-py==0.18.1
scipy==1.14.0
Send2Trash==1.8.3
setuptools==70.1.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
stack-data==0.6.3
tabula-py==2.9.3
tabulate==0.9.0
terminado==0.18.1
thefuzz==0.22.1
tinycss2==1.3.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
Unidecode==1.3.8
UpSetPlot==0.9.0
uri-template==1.3.0
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==0.16.1
wheel==0.43.0
widgetsnbextension==4.0.11
yarl==1.9.4
